Created the room_description_selection
The room_description_selection function looks at the items the player
has in their inventory and checks to see if they match the required
items for the room. If they do not the default description is
displayed. If they do then the second description is displayed, the
default is removed and the players inventory is cleared.